Interactive plots for Rsim

Interactive plots for Rsim plot using plotly:: package and tidyverse::

add another paragraph

Usage

webplot_rpathviz(
  Rpath.obj,
  eco.name = attr(Rpath.obj, "eco.name"),
  line.col = "grey",
  h_spacing = 3,
  node_size_min = 1,
  node_size_max = 30,
  fleet_color = "#B40F20",
  groups_palette = c("#EC7604", "#CB7A5C", "#5785C1", "#0B775E"),
  text_size = 3
)

Arguments

Rpath.obj

An object of class Rpath rpath().

eco.name

Optional name of the ecosystem. Default is the eco.name attribute from the rpath object created from running rpath().

line.col

description.

h_spacing

horizontal spacing multiplier, it spreads nodes horizontally.

node_size_min

from scale_value() Minimum value of the new scale (default is 1).

node_size_max

from scale_value() Maximum value of the new scale (default is 30).

fleet_color

single color for fleet nodes.

groups_palette

color palette for non-fleet groups. Default is a vector of colors.

text_size

size of the text labels.

Value

Returns a plot visualization of the food web.

Contributors

webplot original function from Rpath by Kerim Aydin

Plotting

Caution: For large food webs see the examples.

Examples

if (FALSE) { # \dontrun{
# Read in Rpath parameter file, generate and name model object
Rpath.obj <- Rpath::rpath(Rpath::AB.params, eco.name = "Anchovy Bay")
# Plot food web diagram with all groups labeled, including fleets
webplot_rpathviz(Rpath.obj, h_spacing = 3, text_size = 3)

# Read in Rpath parameter file, generate and name model object
Rpath.obj <- Rpath::rpath(Rpath::Ecosense.EBS, eco.name = "Eastern Bering Sea")
# Plot food web diagram with all groups labeled, including fleets. Follow the steps
# 1) assign the plot to an object
# 2) use ggplot2::ggsave() to save the plot for a fast visualization
wp <- webplot_rpathviz(Rpath.obj, h_spacing = 3, text_size = 3)
ggplot2::ggsave("figures/EBSfoodwebplot2.png", p , width= 16, height= 10)

} # }